requirements.txt
mlflow==3.9.0
cloudpickle==3.1.2
numpy==2.4.2
pandas==2.3.3
psutil==7.2.2
pyarrow==22.0.0
scikit-learn==1.8.0
scipy==1.17.0
pytest==7.3.1
pytest-cov==4.0.0
dvc
PyYAML
joblib
nltk